+++ app/src/main/java/campus/tech/kakao/map/presentation/view/KakaoMapViewActivity.kt
@@ -1,12 +1,18 @@
 package campus.tech.kakao.map.presentation.view
 
+import android.app.AlertDialog
 import android.content.Intent
+import android.content.pm.PackageManager
 import android.graphics.Color
+import android.net.Uri
 import android.os.Bundle
+import android.provider.Settings
 import android.util.Log
 import android.view.View
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
 import androidx.databinding.DataBindingUtil
 import campus.tech.kakao.map.R
 import campus.tech.kakao.map.databinding.ActivityKakaoMapViewBinding
@@ -29,6 +35,14 @@ class KakaoMapViewActivity : AppCompatActivity() {
     private lateinit var binding: ActivityKakaoMapViewBinding
     private var kakaoMap: KakaoMap? = null
     private val viewModel: KakaoMapViewModel by viewModels()
+    private val requestPermissionLauncher = registerForActivityResult(
+        ActivityResultContracts.RequestPermission(),
+    ) { isGranted: Boolean ->
+        if (isGranted)
+            Log.d("Notification Permission", "Notification permission granted.")
+        else
+            Log.d("Notification Permission", "Notification permission denied.")
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -39,6 +53,7 @@ class KakaoMapViewActivity : AppCompatActivity() {
 
         binding.persistentBottomSheet.visibility = View.GONE
 
+        askNotificationPermission()
         initKakaoMap()
         clickSearchButton()
     }
@@ -109,6 +124,41 @@ class KakaoMapViewActivity : AppCompatActivity() {
         )
         val cameraUpdate = CameraUpdateFactory.newCenterPosition(position)
         map.moveCamera(cameraUpdate, CameraAnimation.from(500, true, true))
+    }
+
+    private fun askNotificationPermission() {
+        if (ContextCompat.checkSelfPermission(
+                this,
+                android.Manifest.permission.POST_NOTIFICATIONS
+            ) ==
+            PackageManager.PERMISSION_GRANTED
+        ) {
+            Log.d("SplashScreenActivity", "Notification permission already granted.")
+        } else if (shouldShowRequestPermissionRationale(android.Manifest.permission.POST_NOTIFICATIONS)) {
+            showNotificationPermissionDialog()
+        } else {
+            requestPermissionLauncher.launch(android.Manifest.permission.POST_NOTIFICATIONS)
+        }
+    }
+
+    private fun showNotificationPermissionDialog() {
+        AlertDialog.Builder(this@KakaoMapViewActivity).apply {
+            setTitle(getString(R.string.ask_notification_permission_dialog_title))
+            setMessage(
+                String.format(
+                    "다양한 알림 소식을 받기 위해 권한을 허용하시겠어요?\n(설정에서 %s의 알림 권한을 허용해주세요.",
+                    getString(R.string.app_name)
+                )
+            )
+            setPositiveButton(getString(R.string.accept_notification_permission)) { _, _ ->
+                val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
+                val uri = Uri.fromParts("package", packageName, null)
+                intent.data = uri
+                startActivity(intent)
+            }
+            setNegativeButton(getString(R.string.deny_notification_permission)) { _, _ -> }
+            show()
+        }
     }
 
     override fun onResume() {
